+++ redwood-treehouse-host-composeui/api/jvm/redwood-treehouse-host-composeui.api
@@ -1,4 +1,18 @@
+public class app/cash/redwood/treehouse/composeui/ComposeUiRoot : app/cash/redwood/widget/RedwoodView$Root {
+	public static final field $stable I
+	public fun <init> ()V
+	public fun Render (Landroidx/compose/runtime/Composer;I)V
+	public fun contentState (IZLjava/lang/Throwable;)V
+	public synthetic fun getChildren ()Lapp/cash/redwood/widget/Widget$Children;
+	public fun getChildren ()Lapp/cash/redwood/widget/compose/ComposeWidgetChildren;
+	public fun getModifier ()Lapp/cash/redwood/Modifier;
+	public synthetic fun getValue ()Ljava/lang/Object;
+	public fun getValue ()Lkotlin/jvm/functions/Function2;
+	public fun restart (Lkotlin/jvm/functions/Function0;)V
+	public fun setModifier (Lapp/cash/redwood/Modifier;)V
+}
+
 public final class app/cash/redwood/treehouse/composeui/TreehouseContentKt {
-	public static final fun TreehouseContent (Lapp/cash/redwood/treehouse/TreehouseApp;Lapp/cash/redwood/treehouse/TreehouseView$WidgetSystem;Lapp/cash/redwood/treehouse/TreehouseContentSource;Landroidx/compose/ui/Modifier;Landroidx/compose/runtime/Composer;II)V
+	public static final fun TreehouseContent (Lapp/cash/redwood/treehouse/TreehouseApp;Lapp/cash/redwood/treehouse/TreehouseView$WidgetSystem;Lapp/cash/redwood/treehouse/TreehouseContentSource;Landroidx/compose/ui/Modifier;Lkotlin/jvm/functions/Function1;Landroidx/compose/runtime/Composer;II)V
 }
