+++ samples/emoji-search/android-composeui/src/main/kotlin/com/example/redwood/emojisearch/android/composeui/EmojiSearchActivity.kt
@@ -116,6 +116,7 @@ class EmojiSearchActivity : ComponentActivity() {
             widgetSystem = widgetSystem,
             contentSource = treehouseContentSource,
             modifier = Modifier.padding(contentPadding),
+            root = { scope -> EmojiSearchComposeUiRoot(scope) },
           )
         }
       }
